+++ base/decorators.py
@@ -1,3 +1,9 @@
+from functools import wraps
+from http import HTTPStatus
+import uweb3
+from base.model import model
+
+
 def NotExistsErrorCatcher(f):
   """Decorator to return a 404 if a NotExistError exception was returned."""
 
@@ -8,3 +14,56 @@ def NotExistsErrorCatcher(f):
       return args[0].RequestInvalidcommand(error=error)
 
   return wrapper
+
+
+def apiuser(f):
+  """Decorator to check if the given API key is allowed to access the resource."""
+
+  def wrapper(*args, **kwargs):
+    # This is bypassed if a user is already logged in trough a session
+    if args[0].user:
+      args[0].apikey = None
+      return f(*args, **kwargs)
+    key = None
+    if 'apikey' in args[0].get:
+      key = args[0].get.getfirst('apikey')
+    elif 'apikey' in args[0].post:
+      key = args[0].post.getfirst('apikey')
+    elif 'apikey' in args[0].req.headers:
+      key = args[0].req.headers.get('apikey')
+    try:
+      args[0].apikey = model.Apiuser.FromKey(args[0].connection, key)
+    except model.Apiuser.NotExistError as apierror:
+      return uweb3.Response(content={'error': str(apierror)}, httpcode=403)
+    return f(*args, **kwargs)
+
+  return wrapper
+
+
+def json_error_wrapper(func):
+  wraps(func)
+
+  def wrapper_schema_validation(*args, **kwargs):
+    try:
+      return func(*args, **kwargs)
+    except model.NotExistError as msg:
+      return uweb3.Response(
+          {
+              "error": True,
+              "errors": msg.args,
+              "http_status": HTTPStatus.NOT_FOUND,
+          },
+          httpcode=HTTPStatus.NOT_FOUND,
+      )
+    except Exception as err:
+      print(err)
+      return uweb3.Response(
+          {
+              "error": True,
+              "errors": ["Something went wrong!"],
+              "http_status": HTTPStatus.INTERNAL_SERVER_ERROR,
+          },
+          httpcode=HTTPStatus.INTERNAL_SERVER_ERROR,
+      )
+
+  return wrapper_schema_validation
